# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Founders Finance API
servers:
  - url: /api
    description: Base API path
...
security:
  - sessionCookie: []
paths:
...
  /statements:
    get:
      operationId: listStatements
      tags: [statements]
      summary: List statements
      parameters:
        - name: account_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: include_archived
          in: query
          required: false
          schema:
            type: boolean
          description: Include archived statements in the history view.
      responses:
        "200":
          description: List of statements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Statement"
    post:
      operationId: createStatement
      tags: [statements]
      summary: Create statement metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStatementBody"
      responses:
        "201":
          description: Created statement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Statement"
  /statements/{id}:
    get:
      operationId: getStatement
      tags: [statements]
      summary: Get statement with lines
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Statement detail with lines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatementDetail"
        "404":
          description: Not found
    delete:
      operationId: archiveStatement
      tags: [statements]
      summary: Archive statement while retaining lines and reconciliation history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                type: object
                properties:
                  archived:
                    type: boolean
                  id:
                    type: string
                    format: uuid
                  archived_at:
                    type: string
                    format: date-time
                required:
                  - archived
                  - id
                  - archived_at
  /statements/{id}/lines:
    get:
      operationId: listStatementLines
      tags: [statements]
      summary: List lines for a statement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of statement lines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatementLine"
    post:
      operationId: addStatementLines
      tags: [statements]
      summary: Add lines to statement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddStatementLinesBody"
      responses:
        "200":
          description: Updated statement with lines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatementDetail"
  /statement-lines/{id}/match:
    post:
      operationId: matchStatementLine
      tags: [statements]
      summary: Match statement line to transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchStatementLineBody"
      responses:
        "200":
          description: Updated statement line
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatementLine"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ff_session
...
  schemas:
...
    Statement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        account_name:
          type: string
          nullable: true
        statement_month:
          type: string
          format: date
        opening_balance:
          type: number
          nullable: true
        closing_balance:
          type: number
          nullable: true
        status:
          type: string
          enum:
            - uploaded
            - reconciling
            - reconciled
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        archived_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - account_id
        - statement_month
        - status
        - created_at
        - updated_at
    StatementLine:
      type: object
      properties:
        id:
          type: string
          format: uuid
        statement_id:
          type: string
          format: uuid
        transaction_date:
          type: string
          format: date
          nullable: true
        posted_date:
          type: string
          format: date
          nullable: true
        description:
          type: string
          nullable: true
        amount:
          type: number
        balance_after:
          type: number
          nullable: true
        matched_transaction_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum:
            - unmatched
            - matched
            - ignored
            - needs_review
        notes:
          type: string
          nullable: true
      required:
        - id
        - statement_id
        - amount
        - status
    StatementDetail:
      type: object
      properties:
        statement:
          $ref: "#/components/schemas/Statement"
        lines:
          type: array
          items:
            $ref: "#/components/schemas/StatementLine"
      required:
        - statement
        - lines
    CreateStatementBody:
      type: object
      properties:
        account_id:
          type: string
          format: uuid
        statement_month:
          type: string
          format: date
        opening_balance:
          type: number
          nullable: true
        closing_balance:
          type: number
          nullable: true
      required:
        - account_id
        - statement_month
    AddStatementLinesBody:
      type: object
      properties:
        lines:
          type: array
          items:
            $ref: "#/components/schemas/StatementLineInput"
      required:
        - lines
    StatementLineInput:
      type: object
      properties:
        transaction_date:
          type: string
          format: date
          nullable: true
        posted_date:
          type: string
          format: date
          nullable: true
        description:
          type: string
          nullable: true
        amount:
          type: number
        balance_after:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - amount
...
    MatchStatementLineBody:
      type: object
      properties:
        transaction_id:
          type: string
          format: uuid
        match_type:
          type: string
          default: manual
      required:
        - transaction_id
...